Gradientattempt

See merge request payload_software/Tegra-SFM!67

Former-commit-id: a4aad9f9a156e867135c433eb4027335213eef0b

+++ src/SFM.cu
@@ -100,7 +100,7 @@ int main(int argc, char *argv[]){
     ssrlcv::PointCloudFactory demPoints = ssrlcv::PointCloudFactory();
 
     // bunlde adjustment loop would be here. images_vec woudl be modified to minimize the boi
-    unsigned long long int* linearError = (unsigned long long int*)malloc(sizeof(unsigned long long int));
+    float* linearError = (float*)malloc(sizeof(float));
     ssrlcv::BundleSet bundleSet = demPoints.generateBundles(&matchSet,images);
 
     // the version that will be used normally
